Preliminary work on pixel physics
An attempt to make a simple pixel-based 2d physics engine in pygame.
Have presently managed to draw sprites and do some simple updates on the
display. Working on detecting and resolving collisions. The math is
fairly clear, but finding normal vectors will be a challenge.
Would also like to be able to scale the display. Single pixels are very
small on my current display.