Normal vectors using a normal map.
A sobel transform is used to identify the edges of the terrain, and
assign a normal vector to each coordinate. Mask overlap is used to find
coordinates where terrain boundary intersects object body. This gives a
line of intersection points, and the normal vector is calculated to be
the average of these points' normal vectors. Increase accuracy by
finding TOI more precisely. Currently also uses old collision detection
algorithm to calculate depth of impact. Should rewrite this.